refactor(theme-builder): styles to use docs tailwind
