+++ website/src/pages/themes/_components/slider.tsx
@@ -34,7 +34,7 @@ const Slider = ({
     <div className={className}>
       <label
         htmlFor={id}
-        className="block mb-1 text-sm font-medium text-gray-900 dark:text-white"
+        className="block mb-1 text-sm font-medium text-grayscale-900 dark:text-white"
       >
         <span className="text-sm font-bold">{label}</span>: {value}
         {unit}
@@ -44,7 +44,7 @@ const Slider = ({
         type="range"
         value={value}
         onChange={handleChange}
-        className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer dark:bg-gray-700"
+        className="w-full h-2 bg-grayscale-300 rounded-lg appearance-none cursor-pointer dark:bg-grayscale-700"
         min={min}
         max={max}
         step={step}
