--- website/src/pages/themes/_components/tailwind.css
@@ -1,9 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer base {
-  html {
-    font-family: "Roboto", system-ui, sans-serif;
-  }
-}
